refactor: reorganize simulation logic and update circuit handling
