+++ src/lib/simulation/logic.ts
@@ -1,4 +1,4 @@
-import { CircuitState } from "@/types/enums";
+import { CircuitState, GateMetadata, Operation } from "@/types/enums";
 
 const mask = 1;
 
@@ -55,3 +55,40 @@ export const Logic = {
     return Logic.not([val]);
   },
 };
+
+export function computeGateSignal(
+  op: Operation,
+  inputs: CircuitState[]
+): CircuitState {
+  const gateMetadata = GateMetadata[op];
+  if (!gateMetadata) {
+    return CircuitState.FLOATING;
+  }
+  const { inputs: requiredInputs } = gateMetadata;
+  if (inputs.length < requiredInputs) return CircuitState.FLOATING;
+
+  switch (op) {
+    case Operation.ConstantLow:
+      return CircuitState.LOW;
+    case Operation.ConstantHigh:
+      return CircuitState.HIGH;
+    case Operation.Output:
+      return inputs[0] ?? CircuitState.LOW;
+    case Operation.And:
+      return Logic.and(inputs);
+    case Operation.Or:
+      return Logic.or(inputs);
+    case Operation.Not:
+      return Logic.not(inputs);
+    case Operation.Nand:
+      return Logic.nand(inputs);
+    case Operation.Nor:
+      return Logic.nor(inputs);
+    case Operation.Xor:
+      return Logic.xor(inputs);
+    // case Operation.Xnor:
+    //   return Logic.xnor(inputs);
+    default:
+      return CircuitState.FLOATING;
+  }
+}
